requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.6.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
starlette-session==0.4.0
jinja2==3.1.4
python-multipart==0.0.20
anthropic==0.40.0
openai==1.57.0
python-dotenv==1.0.1
aiofiles==24.1.0
